Adding Google API using a fragemnt
The fragment is blocking the RelativeLayout so I need to make it smaller. I also need to plan how the maps will load the locations on the wheel. I also added a new music .raw file but there is still "popping" going on. And changing the AVD doesn't work either.